quant-platform/services/trading-agents-team/requirements.txt
anthropic==0.34.2
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
redis==5.0.8
APScheduler==3.10.4
python-dotenv==1.0.1
feedparser==6.0.11
httpx==0.27.2
pydantic==2.9.2
yfinance==0.2.44
finance-datareader
rank-bm25
